+++ lib/v2.0/C++/pprzlink/IvyLink.h
@@ -79,7 +79,7 @@ namespace pprzlink {
      *
      * @param bindId
      */
-    void UnbindMessage(int bindId);
+    void UnbindMessage(long bindId);
 
     /**
      *
